refactor: extension 코드 분리
